my girlfriend shouts at me that I need to leave.
