+++ crates/typst-cli/Cargo.toml
@@ -55,6 +55,7 @@ tempfile = { workspace = true }
 ureq = { workspace = true }
 xz2 = { workspace = true, optional = true }
 zip = { workspace = true, optional = true }
+url = "2.5.0"
 
 [build-dependencies]
 clap = { workspace = true, features = ["string"] }
