+++ crates/typst-cli/src/world.rs
@@ -7,9 +7,10 @@ use chrono::{DateTime, Datelike, Local};
 use comemo::Prehashed;
 use ecow::eco_format;
 use parking_lot::Mutex;
-use typst::diag::{FileError, FileResult, StrResult};
+use url::Url;
+use typst::diag::{FileError, FileResult, PackageResult, StrResult};
 use typst::foundations::{Bytes, Datetime, Dict, IntoValue};
-use typst::syntax::{FileId, Source, VirtualPath};
+use typst::syntax::{FileId, Source, VirtualPath, REMOTE_PACKAGE};
 use typst::text::{Font, FontBook};
 use typst::{Library, World};
 use typst_timing::{timed, TimingScope};
@@ -18,6 +19,7 @@ use crate::args::SharedArgs;
 use crate::compile::ExportCache;
 use crate::fonts::{FontSearcher, FontSlot};
 use crate::package::prepare_package;
+use crate::remote::HTTPRemoteAssetFetcher;
 
 /// A world that provides access to the operating system.
 pub struct SystemWorld {
@@ -43,6 +45,8 @@ pub struct SystemWorld {
     /// The export cache, used for caching output files in `typst watch`
     /// sessions.
     export_cache: ExportCache,
+
+    fetcher: HTTPRemoteAssetFetcher,
 }
 
 impl SystemWorld {
@@ -82,6 +86,8 @@ impl SystemWorld {
 
             Library::builder().with_inputs(inputs).build()
         };
+        let pb = PathBuf::from("/tmp/some-shit");
+        let fetcher = HTTPRemoteAssetFetcher::new(pb);
 
         Ok(Self {
             workdir: std::env::current_dir().ok(),
@@ -94,6 +100,7 @@ impl SystemWorld {
             slots: Mutex::new(HashMap::new()),
             now: OnceLock::new(),
             export_cache: ExportCache::new(),
+            fetcher,
         })
     }
 
@@ -118,7 +125,7 @@ impl SystemWorld {
             .get_mut()
             .values()
             .filter(|slot| slot.accessed())
-            .filter_map(|slot| system_path(&self.root, slot.id).ok())
+            .filter_map(|slot| system_path(&self.root, slot.id, &self.fetcher).ok())
     }
 
     /// Reset the compilation state in preparation of a new compilation.
@@ -160,11 +167,11 @@ impl World for SystemWorld {
     }
 
     fn source(&self, id: FileId) -> FileResult<Source> {
-        self.slot(id, |slot| slot.source(&self.root))
+        self.slot(id, |slot| slot.source(&self.root, &self.fetcher))
     }
 
     fn file(&self, id: FileId) -> FileResult<Bytes> {
-        self.slot(id, |slot| slot.file(&self.root))
+        self.slot(id, |slot| slot.file(&self.root, &self.fetcher))
     }
 
     fn font(&self, index: usize) -> Option<Font> {
@@ -229,9 +236,9 @@ impl FileSlot {
     }
 
     /// Retrieve the source for this file.
-    fn source(&mut self, project_root: &Path) -> FileResult<Source> {
+    fn source(&mut self, project_root: &Path, fetcher: &HTTPRemoteAssetFetcher) -> FileResult<Source> {
         self.source.get_or_init(
-            || system_path(project_root, self.id),
+            || system_path(project_root, self.id, fetcher),
             |data, prev| {
                 let name = if prev.is_some() { "reparsing file" } else { "parsing file" };
                 let _scope = TimingScope::new(name, None);
@@ -247,9 +254,9 @@ impl FileSlot {
     }
 
     /// Retrieve the file's bytes.
-    fn file(&mut self, project_root: &Path) -> FileResult<Bytes> {
+    fn file(&mut self, project_root: &Path, fetcher: &HTTPRemoteAssetFetcher) -> FileResult<Bytes> {
         self.file
-            .get_or_init(|| system_path(project_root, self.id), |data, _| Ok(data.into()))
+            .get_or_init(|| system_path(project_root, self.id, fetcher), |data, _| Ok(data.into()))
     }
 }
 
@@ -313,10 +320,25 @@ impl<T: Clone> SlotCell<T> {
 }
 
 
+fn prepare_package_or_remote_file(id: FileId, fetcher: &HTTPRemoteAssetFetcher) -> PackageResult<Option<PathBuf>>{
+    if id.package().is_none() {
+        return Ok(None)
+    }
+    let spec = id.package().unwrap();
+    if spec == &REMOTE_PACKAGE {
+        let url_str = id.vpath().as_rootless_path().to_str().unwrap();
+        println!("here is the url: {}", url_str);
+        let url = Url::parse(url_str).unwrap();
+        fetcher.fetch(&url).unwrap();
+        return Ok(Some(fetcher.mirror_root()));
+    }
+    prepare_package(spec).map(|x| Some(x))
+}
+
 // todo: download file if not present.
 /// Resolves the path of a file id on the system, downloading a package if
 /// necessary.
-fn system_path(project_root: &Path, id: FileId) -> FileResult<PathBuf> {
+fn system_path_old(project_root: &Path, id: FileId) -> FileResult<PathBuf> {
     // Determine the root path relative to which the file path
     // will be resolved.
     let buf;
@@ -329,6 +351,24 @@ fn system_path(project_root: &Path, id: FileId) -> FileResult<PathBuf> {
     // Join the path to the root. If it tries to escape, deny
     // access. Note: It can still escape via symlinks.
     id.vpath().resolve(root).ok_or(FileError::AccessDenied)
+}
+
+fn system_path(project_root: &Path, id: FileId, fetcher: &HTTPRemoteAssetFetcher) -> FileResult<PathBuf> {
+    // println!("fuck me");
+    // Determine the root path relative to which the file path
+    // will be resolved.
+    // let p = PathBuf::from("/tmp/hello");
+    // let fetcher = HTTPRemoteAssetFetcher::new(p);
+    let root = prepare_package_or_remote_file(id, fetcher)?
+        .unwrap_or(PathBuf::from(project_root));
+    // if let Some(spec) = id.package() {
+    //     buf = prepare_package(spec)?;
+    //     root = &buf;
+    // }
+
+    // Join the path to the root. If it tries to escape, deny
+    // access. Note: It can still escape via symlinks.
+    id.vpath().resolve(root.as_path()).ok_or(FileError::AccessDenied)
 }
 
 /// Read a file.
